simplify balance sheet
